ml-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
chronos-forecasting==1.3.1
torch==2.4.0
numpy==1.26.4
pydantic==2.8.0
